requirements.txt
Django==1.11.3
django-authtools
django-csvimport
django-material
psycopg2